fix(clippy): resolve all -D warnings - sort_by_key + remove redundant casts
- app_manager.rs: sort_by(|a,b| b.0.cmp(&a.0)) -> sort_by_key(|b| Reverse(b.0))
- windows.rs: VK_RETURN.0 as u16 / VK_SPACE.0 as u16 -> .0 (u16 already)

cargo clippy -- -D warnings now exits clean.

+++ src-tauri/src/platform/windows.rs
@@ -119,8 +119,8 @@ pub fn simulate_click(button: &str, count: u32) -> Result<(), String> {
 
 pub fn simulate_key(key: &str, _modifiers: &[String]) -> Result<(), String> {
     let vk = match key {
-        "Enter" | "Return" => VK_RETURN.0 as u16,
-        "Space" => VK_SPACE.0 as u16,
+        "Enter" | "Return" => VK_RETURN.0,
+        "Space" => VK_SPACE.0,
         _ => return Err(format!("unsupported key: '{key}'")),
     };
     send_key_input(vk, false)?;
